core: canonicalize call attribution across protocols

+++ src/core/util/call_state_internal.h
@@ -12,7 +12,7 @@ typedef struct {
     uint64_t epoch;
     uint8_t ended_committed;
     uint64_t notice_epoch;
-    uint32_t notice_target_id;
+    uint64_t notice_target_id;
     uint8_t notice_kind;
     uint8_t notice_handled;
 } dsd_call_event_lifecycle;
